feat: ensure product quantity can be decrease

+++ src/actions/decrease-cart-product-quantity/index.ts
@@ -8,12 +8,12 @@ import { db } from "@/db";
 import { cartItemTable } from "@/db/schema";
 import { auth } from "@/lib/auth";
 
-import { removeCartProductsSchema } from "./schema";
+import { decreaseCartProductQuantitySchema } from "./schema";
 
-export const removeCartProducts = async (
-  data: z.infer<typeof removeCartProductsSchema>,
+export const decreaseCartProductQuantity = async (
+  data: z.infer<typeof decreaseCartProductQuantitySchema>,
 ) => {
-  removeCartProductsSchema.parse(data);
+  decreaseCartProductQuantitySchema.parse(data);
 
   const session = await auth.api.getSession({
     headers: await headers(),
@@ -36,6 +36,10 @@ export const removeCartProducts = async (
   if (cartDoesNotBelongToUser) {
     throw new Error("Unauthorized");
   }
-
-  await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+  if (cartItem.quantity === 1) {
+    await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+    return;
+  }
+  await db.update(cartItemTable).set({ quantity: cartItem.quantity - 1 }).where(eq(cartItemTable.id, cartItem.id));
+  
 };
